requirements.txt
altgraph===0.17.4
bcrypt===4.1.3
cffi===1.16.0
cryptography===42.0.8
packaging===24.1
paramiko===3.4.0
pefile===2023.2.7
pycparser===2.22
pyinstaller===6.9.0
pyinstaller-hooks-contrib===2024.7
PyNaCl===1.5.0
python-dotenv===1.0.1
pywin32-ctypes===0.2.2
scp===0.15.0